refactor some of the transaction hooks code to use selectors more effectively
